Revert "switch to litehtml"
This reverts commit 26c782e20481be939c2269c0db42ba3c93cc1bc8.

+++ src/screenshot.cpp
@@ -58,7 +58,7 @@ namespace ocr {
 	HBITMAP ScreenshotWnd::captureEntireScreen() {
 		const auto [width, height] = getScreenSize();
 
-		return captureScreenRegion(cv::Rect(0, 0, width, height));
+		return captureScreenRegion(cv::Rect(0, 0, static_cast<int>(width), static_cast<int>((height))));
 	}
 
 	HBITMAP ScreenshotWnd::captureScreenRegion(const cv::Rect rect) {
